OSSL_STORE_CACHE: new functionality to cache the objects retrieved
A OSSL_STORE_CACHE can cache objects for a number of URIs.  The idea
is to have the cache as an intermediary loader that saves away the
objects that are loaded "on the fly", and being able to return them
immediately later on when using that same cache.

For now, the cache is rather simple, without any aging or other
techniques usually associated with a cache, but that can certainly be
expanded in the future.